requirements.txt
setuptools
python-dotenv
Flask
Pyrebase4
firebase_admin
flask_login
functools